modified testing files, added config.py file
Modified testing files and added config.py file required for testing
